+++ system/core/URI.php
@@ -628,9 +628,7 @@ class CI_URI {
 	 */
 	public function ruri_string()
 	{
-		global $RTR;
-
-		return ltrim($RTR->directory, '/').implode('/', $this->rsegments);
+		return ltrim(load_class('Router', 'core')->directory, '/').implode('/', $this->rsegments);
 	}
 
 }